M5: the mixer e2e suite, the docs sweep, and a key that never worked
tests/e2e/mixer_e2e_test.cpp, the docs M5 changed, and the milestone checkpoint:
Docker CI green on all four presets and clang-tidy clean.

THE GOLDEN COVERS MORE THAN THE PLAN EXPECTED. docs/TESTING.md allows a
committed hash only where the path has no `a*b+c` for an FMA to fuse. Audited by
reading the three places the mixer does arithmetic -- apply_gain is `*=`,
apply_balance is `*=`, mix_into is `+=` -- so a lone multiply and a lone add,
both exactly specified by IEEE-754. The ENTIRE level and routing half is inside
the rule, and the golden exercises strip gain, balance, mute, bus routing and bus
gain rather than settling for a transparent chain. Verified on AppleClang/arm64
and clang-18/x86-64.

The three processors are outside it -- each has a fused expression in its inner
loop -- so engaged they assert block-size invariance and run-to-run equality,
which is also the property they are most able to break.

TWO FINDINGS, BOTH FROM THE TESTS RATHER THAN FROM INSPECTION.

The invariance test failed first with every processor switched off, which a
working mixer cannot do. Two causes, in order:

  1. `kOneBlock` was the whole render, 12000 frames against a 2048-frame
     ceiling. Engine::render() asserts that bound and the dev preset is NDEBUG,
     so the assertion was compiled out and the render wrote past every graph
     buffer. Running the same case under asan -- which keeps assertions --
     aborted on it immediately. A golden computed on dev alone would have been a
     hash of a buffer overrun, and perfectly reproducible.
  2. The hits were staggered with PadEvent::frame_offset, which is defined as an
     offset from the start of the BLOCK and clamped to num_frames - 1. So 259
     lands at frame 259 in a 2048-frame block and at 63 in a 64-frame one.
     That is the contract, not a hole in it: nothing real produces such an
     offset. Both are now in docs/TESTING.md.

AND PERFORM'S TAB HAS NEVER WORKED. Carried from T8 rather than filed: the
right-hand panel switch has been bound to Tab since M2, and FTXUI consumes Tab
before CatchEvent, so it never reached the program. It is now on `\` as well --
Tab is kept because it costs nothing and starts working the day the Kitty
protocol is negotiated. The PTY assertion that should have caught it was guarded
by an `if` that made it vacuous; it presses the key unconditionally now, and
removing the binding fails the session.

clang-tidy found 18 sites, one of them worth the name: a use-after-move in
strip_test.cpp that turned out to be clang-tidy reading a Catch2 macro expansion
rather than a real defect. The rest were style, an integer division feeding a
float conversion, and three float loop counters now stepped by an integer index
-- which is more accurate as well as quieter.

Docs: ARCHITECTURE gains the mixer graph as built and stops calling it unbuilt;
TESTING gains the M5 acceptance table and the two findings above; README moves
from M3 to M5 and documents the mixer verbs; MIXER.md was already current.

483 tests on dev/asan/ubsan, 482 on tsan, on macOS and in Docker.

+++ tests/unit/engine_render_test.cpp
@@ -1446,7 +1446,9 @@ std::shared_ptr<const rt::PadConfig> graph_config(std::uint8_t pad) {
     for (std::size_t frame = 0; frame < data.size(); ++frame) {
       // Irrational-looking but integer-derived: no libm call whose last bit
       // could differ between platforms, and no repeating value.
-      const auto mixed = static_cast<float>(((frame * 37) + (pad * 101) + (channel * 7)) % 9'973);
+      const auto mixed = static_cast<float>(((frame * 37) + (static_cast<std::size_t>(pad) * 101) +
+                                             (static_cast<std::size_t>(channel) * 7)) %
+                                            9'973);
       data[frame] = ((mixed / 4'986.5F) - 1.0F) * 0.31F;
     }
   }
